ai-image-caption-service/requirements.txt
Flask==3.1.2
Flask-Cors
easyocr
pillow
openai
python-dotenv